Signers is now a package
