+++ nucypher/blockchain/eth/__init__.py
@@ -14,17 +14,9 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
+
 from pathlib import Path
 
-from nucypher.blockchain.eth.signers import Signer, ClefSigner, KeystoreSigner, TrezorSigner
-
-SIGNERS = {
-    ClefSigner.URI_SCHEME: ClefSigner,
-    KeystoreSigner.URI_SCHEME: KeystoreSigner,
-    TrezorSigner.URI_SCHEME: TrezorSigner,
-}
-
-Signer.SIGNERS = SIGNERS
-
+# TODO: Move to constants.py?
 BASE_DIRECTORY = Path(__file__).parent
 CONTRACT_REGISTRY_BASE = BASE_DIRECTORY / "contract_registry"
